+++ packages/realm/src/app-services/App.ts
@@ -16,6 +16,7 @@
 //
 ////////////////////////////////////////////////////////////////////////////
 
+import { inject } from "src/platform/network";
 import {
   AnyUser,
   Credentials,
@@ -30,6 +31,7 @@ import {
   createNetworkTransport,
   deviceInfo,
   fs,
+  network,
 } from "../internal";
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -140,6 +142,12 @@ export type AppConfiguration = {
    * @since 12.2.0
    */
   metadata?: Metadata;
+
+  /**
+   * Specify a custom `fetch` implementation.
+   * @since 12.3.0
+   */
+  fetchOverride?: any;
 };
 
 /**
@@ -261,7 +269,8 @@ export class App<
   constructor(configOrId: AppConfiguration | string) {
     const config: AppConfiguration = typeof configOrId === "string" ? { id: configOrId } : configOrId;
     assert.object(config, "config");
-    const { id, baseUrl, app, timeout, multiplexSessions = true, baseFilePath, metadata } = config;
+    const { id, baseUrl, app, timeout, multiplexSessions = true, baseFilePath, metadata, fetchOverride } = config;
+
     assert.string(id, "id");
     if (timeout !== undefined) {
       assert.number(timeout, "timeout");
@@ -275,6 +284,10 @@ export class App<
       if (metadata.mode === MetadataMode.Encryption) {
         assert(metadata.encryptionKey, "encryptionKey is required");
       }
+    }
+
+    if (fetchOverride) {
+      inject({ fetch: fetchOverride });
     }
 
     fs.ensureDirectoryForFile(fs.joinPaths(baseFilePath || fs.getDefaultDirectoryPath(), "mongodb-realm"));